test/key_gen: unit test for group_encrypt
